fix: filters not applied in helm app list
